Moving apply functions to correspondence generators
